Separate aiohttp tests into a file that runs only on py3
