+++ tests/unit/test_clients.py
@@ -6,48 +6,11 @@ import pytest
 
 # Local imports
 from uplink.clients import (
-    AiohttpClient,
     interfaces,
     requests_,
     twisted_,
     register,
     io,
-)
-
-try:
-    from uplink.clients import aiohttp_
-except (ImportError, SyntaxError):
-    aiohttp_ = None
-
-try:
-    import asyncio
-except (ImportError):
-    asyncio = None
-
-if asyncio:
-
-    class AsyncMock(object):
-        def __init__(self, result=None):
-            self._result = result
-            self._calls = 0
-
-        @asyncio.coroutine
-        def __call__(self, *args, **kwargs):
-            self._calls += 1
-            f = asyncio.Future()
-            f.set_result(self._result)
-            return f
-
-        @property
-        def called(self):
-            return self._calls > 0
-
-else:
-    AsyncMock = None
-
-
-requires_python34 = pytest.mark.skipif(
-    not aiohttp_, reason="Requires Python 3.4 or above"
 )
 
 
@@ -221,256 +184,3 @@ class TestTwisted(object):
 
     def test_io(self):
         assert isinstance(twisted_.TwistedClient.io(), io.TwistedStrategy)
-
-
-@pytest.fixture
-def aiohttp_session_mock(mocker):
-    import aiohttp
-
-    return mocker.Mock(spec=aiohttp.ClientSession)
-
-
-class TestAiohttp(object):
-    @requires_python34
-    def test_init_when_aiohttp_is_not_installed(self):
-        with _patch(aiohttp_, "aiohttp", None):
-            with pytest.raises(NotImplementedError):
-                AiohttpClient()
-
-    @requires_python34
-    def test_init_with_session_None(self, mocker):
-        mocker.spy(AiohttpClient, "_create_session")
-        AiohttpClient(kwarg1="value")
-        AiohttpClient._create_session.assert_called_with(kwarg1="value")
-
-    @requires_python34
-    def test_get_client(self, aiohttp_session_mock):
-        client = register.get_client(aiohttp_session_mock)
-        assert isinstance(client, aiohttp_.AiohttpClient)
-
-    @requires_python34
-    def test_request_send(self, mocker, aiohttp_session_mock):
-        # Setup
-        expected_response = mocker.Mock()
-
-        @asyncio.coroutine
-        def request(*args, **kwargs):
-            return expected_response
-
-        aiohttp_session_mock.request = request
-        client = aiohttp_.AiohttpClient(aiohttp_session_mock)
-
-        # Run
-        response = client.send((1, 2, {}))
-        loop = asyncio.get_event_loop()
-        value = loop.run_until_complete(asyncio.ensure_future(response))
-
-        # Verify
-        assert value == expected_response
-
-    @requires_python34
-    def test_callback(self, mocker, aiohttp_session_mock):
-        # Setup
-        expected_response = mocker.Mock(spec=aiohttp_.aiohttp.ClientResponse)
-
-        @asyncio.coroutine
-        def request(*args, **kwargs):
-            return expected_response
-
-        aiohttp_session_mock.request = request
-        client = aiohttp_.AiohttpClient(aiohttp_session_mock)
-        client._sync_callback_adapter = asyncio.coroutine
-
-        # Run
-        globals_ = dict(globals(), client=client)
-        locals_ = {"asyncio": asyncio}
-        exec(
-            """@asyncio.coroutine
-def call():
-    response = yield from client.send((1, 2, {}))
-    response = yield from client.apply_callback(lambda x: 2, response)
-    return response
-""",
-            globals_,
-            locals_,
-        )
-
-        loop = asyncio.get_event_loop()
-        value = loop.run_until_complete(
-            asyncio.ensure_future(locals_["call"]())
-        )
-
-        # Verify
-        assert value == 2
-
-    @requires_python34
-    def test_wrap_callback(self, mocker):
-        # Setup
-        c = AiohttpClient()
-        mocker.spy(c, "_sync_callback_adapter")
-
-        # Run: with callback that is not a coroutine
-        def callback(*_):
-            pass
-
-        c.wrap_callback(callback)
-
-        # Verify: Should wrap it
-        c._sync_callback_adapter.assert_called_with(callback)
-
-        # Run: with coroutine callback
-        coroutine_callback = asyncio.coroutine(callback)
-        assert c.wrap_callback(coroutine_callback) is coroutine_callback
-
-    @requires_python34
-    def test_threaded_callback(self, mocker):
-        def callback(response):
-            return response
-
-        # Mock response.
-        response = mocker.Mock(spec=aiohttp_.aiohttp.ClientResponse)
-        response.text = AsyncMock()
-
-        # Run
-        new_callback = aiohttp_.threaded_callback(callback)
-        return_value = new_callback(response)
-        loop = asyncio.get_event_loop()
-        value = loop.run_until_complete(asyncio.ensure_future(return_value))
-
-        # Verify
-        assert response.text.called
-        assert value == response
-
-        # Run: Verify with callback that returns new value
-        def callback(*_):
-            return 1
-
-        new_callback = aiohttp_.threaded_callback(callback)
-        awaitable = new_callback(response)
-        loop = asyncio.get_event_loop()
-        value = loop.run_until_complete(asyncio.ensure_future(awaitable))
-        assert value == 1
-        assert response.text.called
-
-        # Run: Verify with response that is not ClientResponse (should not be wrapped)
-        response = mocker.Mock()
-        awaitable = new_callback(response)
-        loop = asyncio.get_event_loop()
-        loop.run_until_complete(asyncio.ensure_future(awaitable))
-        assert not response.text.called
-
-    @requires_python34
-    def test_threaded_coroutine(self):
-        @asyncio.coroutine
-        def coroutine():
-            return 1
-
-        threaded_coroutine = aiohttp_.ThreadedCoroutine(coroutine)
-
-        # Run -- should block
-        response = threaded_coroutine()
-
-        # Verify
-        assert response == 1
-
-    @requires_python34
-    def test_threaded_response(self, mocker):
-        @asyncio.coroutine
-        def coroutine():
-            return 1
-
-        def not_a_coroutine():
-            return 2
-
-        response = mocker.Mock()
-        response.coroutine = coroutine
-        response.not_coroutine = not_a_coroutine
-        threaded_response = aiohttp_.ThreadedResponse(response)
-
-        # Run
-        threaded_coroutine = threaded_response.coroutine
-        return_value = threaded_coroutine()
-
-        # Verify
-        assert isinstance(threaded_coroutine, aiohttp_.ThreadedCoroutine)
-        assert return_value == 1
-        assert threaded_response.not_coroutine is not_a_coroutine
-
-    @requires_python34
-    def test_create(self, mocker):
-        session_cls_mock = mocker.patch("aiohttp.ClientSession")
-        positionals = [1]
-        keywords = {"keyword": 2}
-
-        # Run: Create client
-        client = aiohttp_.AiohttpClient.create(*positionals, **keywords)
-
-        # Verify: session hasn't been created yet.
-        assert not session_cls_mock.called
-
-        # Run: Get session
-        loop = asyncio.get_event_loop()
-        loop.run_until_complete(asyncio.ensure_future(client.session()))
-
-        # Verify: session created with args
-        session_cls_mock.assert_called_with(*positionals, **keywords)
-
-    @requires_python34
-    def test_close_auto_created_session(self, mocker):
-        # Setup
-        import gc
-        import aiohttp
-
-        mock_session = mocker.Mock(spec=aiohttp.ClientSession)
-        session_cls_mock = mocker.patch("aiohttp.ClientSession")
-        session_cls_mock.return_value = mock_session
-
-        positionals = [1]
-        keywords = {"keyword": 2}
-
-        # Run: Create client
-        client = aiohttp_.AiohttpClient.create(*positionals, **keywords)
-
-        # Run: Get session
-        loop = asyncio.get_event_loop()
-        loop.run_until_complete(asyncio.ensure_future(client.session()))
-
-        # Verify: session created with args
-        session_cls_mock.assert_called_with(*positionals, **keywords)
-        del client
-        gc.collect()
-        session_cls_mock.return_value.close.assert_called_with()
-
-    @requires_python34
-    def test_exceptions(self):
-        import aiohttp
-
-        exceptions = aiohttp_.AiohttpClient.exceptions
-
-        with pytest.raises(exceptions.BaseClientException):
-            raise aiohttp.ClientError()
-
-        with pytest.raises(exceptions.BaseClientException):
-            # Test polymorphism
-            raise aiohttp.InvalidURL("invalid")
-
-        with pytest.raises(exceptions.ConnectionError):
-            raise aiohttp.ClientConnectionError()
-
-        with pytest.raises(exceptions.ConnectionTimeout):
-            raise aiohttp.ClientConnectorError.__new__(
-                aiohttp.ClientConnectorError
-            )
-
-        with pytest.raises(exceptions.ServerTimeout):
-            raise aiohttp.ServerTimeoutError()
-
-        with pytest.raises(exceptions.SSLError):
-            raise aiohttp.ClientSSLError.__new__(aiohttp.ClientSSLError)
-
-        with pytest.raises(exceptions.InvalidURL):
-            raise aiohttp.InvalidURL("invalid")
-
-    @requires_python34
-    def test_io(self):
-        assert isinstance(aiohttp_.AiohttpClient.io(), io.AsyncioStrategy)
